Float >:(

+++ mywatchlist/models.py
@@ -3,7 +3,7 @@ from django.db import models
 class WatchListItem(models.Model):
     watched = models.BooleanField()
     title = models.CharField(max_length=255)
-    rating = models.FloatField()
+    rating = models.IntegerField()
     release_date = models.TextField()
     review = models.TextField()
